+++ ChangeLog.md
@@ -14,6 +14,7 @@
     * Added utf8-cpp dependency to allow Console to delete multi-byte chars
     * Added stream-style printing to Console
     * Using Console like Stream everywhere, removed old print method
+    * Added Screenshot command back
 
     [ 20140809 ]
     * Script Unit Test brings it’s own scripts to test